refactor: move transaction import API workflow
